add endpoint for saving resource
